+++ toqito/helper/__init__.py
@@ -2,3 +2,4 @@
 from toqito.helper.expr_as_np_array import expr_as_np_array
 from toqito.helper.np_array_as_expr import np_array_as_expr
 from toqito.helper.update_odometer import update_odometer
+from toqito.helper.cvx_kron import cvx_kron